refactor: remove unnecessary fixture
