+++ test/StrategyManager/fixtures/zap.fixture.ts
@@ -4,7 +4,7 @@ import { BigNumber } from '@ryze-blockchain/ethereum'
 import { UniswapV2, UniswapV2ZapHelper, UniswapV3, UniswapV3ZapHelper } from '@src/helpers'
 import { NetworkService } from '@src/NetworkService'
 import { ProjectDeployer } from '@src/ProjectDeployer'
-import { TestERC20__factory, UniswapV2Pair__factory } from '@src/typechain'
+import { TestERC20__factory, UniswapV2Pair__factory, UniswapV3Pool__factory } from '@src/typechain'
 import { parseEther } from 'ethers'
 import hre from 'hardhat'
 
@@ -134,6 +134,16 @@ export async function zapFixture() {
         account0,
     )
 
+    const stableBtcLpUniV3 = UniswapV3Pool__factory.connect(
+        await factoryUniV3.getPool(stablecoin, wbtc, 3000),
+        account0,
+    )
+
+    const btcEthLpUniV3 = UniswapV3Pool__factory.connect(
+        await factoryUniV3.getPool(weth, wbtc, 3000),
+        account0,
+    )
+
     const stableBtcPoolId = await dca.getPoolsLength()
 
     await dca.createPool(
@@ -156,6 +166,8 @@ export async function zapFixture() {
 
     const strategyId = await strategyManager.getStrategiesLength()
     const initialTreasuryBalance = await stablecoin.balanceOf(treasury)
+
+    await liquidityManager.setPositionManagerWhitelist(positionManagerUniV3, true)
 
     return {
         // accounts
@@ -191,6 +203,8 @@ export async function zapFixture() {
         initialTreasuryBalance,
         permitAccount0,
         btcEthLpUniV2,
+        stableBtcLpUniV3,
+        btcEthLpUniV3,
 
         // helpers
         uniswapV2ZapHelper,
